Renamed AudioItem in separated file
